+++ src/index.ts
@@ -2,16 +2,13 @@ import express, { Express } from 'express';
 require('express-async-errors');
 import session from 'express-session';
 import path from 'node:path';
-import multer from 'multer'
 import { tables } from './tables';
-import bodyParser from 'body-parser'
 import { sessionStore } from './dependencies/session';
 import { pageRouter } from './pages';
 import { actionsRouter } from './actions';
 import { registerAuth } from './dependencies/authn';
 import { subscribeAll } from './dependencies/eventStore';
 
-import { uploadedGedcom } from './actions/uploaderGedcom'
 
 const PORT: number = parseInt(process.env.PORT ?? '3000');
 
@@ -57,19 +54,3 @@ app.listen(PORT, (): void => {
   });
 });
 
-const FILE_SIZE_LIMIT_MB = 50
-
-const upload = multer({
-  dest: 'temp',
-  limits: { fileSize: FILE_SIZE_LIMIT_MB * 1024 * 1024 /* MB */ },
-})
-
-app.post(
-  '/importGedcom.html',
-  bodyParser.urlencoded({
-    extended: false,
-    limit: '10mb',
-  }),
-  upload.single('fichier'),
-   uploadedGedcom
-)
