+++ src/pages/home/HomePage.stories.tsx
@@ -7,6 +7,6 @@ export default { title: "Page d'accueil", component: HomePage }
 
 export const Basique = () => (
   <SessionContext.Provider value={{ isLoggedIn: true, userName: 'toto' }}>
-    <HomePage prop1='qsd' />
+    <HomePage />
   </SessionContext.Provider>
 )
