+++ src/pages/home/HomePage.tsx
@@ -2,6 +2,11 @@ import * as React from 'react'
 import { withBrowserBundle } from '../../libs/ssr'
 import { AppLayout } from '../_components/layout/AppLayout'
 
-export const HomePage = withBrowserBundle(() => {
-  return <AppLayout>Bienvenu sur Trésor de famille</AppLayout>
+export const HomePage = (() => {
+  return <AppLayout>Bienvenu sur Trésor de famille 
+    <a href='/importGedcom.html'> GedCom </a>
+  </AppLayout>
+
 })
+
+
