+++ src/pages/index.ts
@@ -1,3 +1,4 @@
 export * from './_components';
 export * from './home';
 export * from './pageRouter';
+export * from './importGedcom'
